requirements.txt
streamlit
firebase-admin

langchain
langchain-openai

python-dateutil
pyyaml

gTTS
